+++ src/frontend/context/Kompetanse/KompetanseContext.tsx
@@ -1,9 +1,9 @@
 import { useState, useEffect } from 'react';
 
-import { sorterEøsPerioder } from '../../komponenter/Fagsak/Behandlingsresultat/Eøs/useEøs';
 import type { IBehandling } from '../../typer/behandling';
 import type { IRestKompetanse } from '../../typer/eøsPerioder';
 import { EøsPeriodeStatus } from '../../typer/eøsPerioder';
+import { sorterEøsPerioder } from '../../utils/eøs';
 
 interface IProps {
     åpenBehandling: IBehandling;
