+++ src/frontend/context/UtenlandskPeriodeBeløp/UtenlandskPeriodeBeløpSkjemaContext.tsx
@@ -6,10 +6,6 @@ import { feil, ok, useFelt, useSkjema } from '@navikt/familie-skjema';
 import type { Ressurs } from '@navikt/familie-typer';
 import { byggTomRessurs, RessursStatus } from '@navikt/familie-typer';
 
-import {
-    konverterDesimalverdiTilSkjemaVisning,
-    konverterSkjemaverdiTilDesimal,
-} from '../../komponenter/Fagsak/Behandlingsresultat/Eøs/useEøs';
 import type { IBehandling } from '../../typer/behandling';
 import type { OptionType } from '../../typer/common';
 import type {
@@ -20,6 +16,10 @@ import type {
 } from '../../typer/eøsPerioder';
 import type { IIsoMånedPeriode } from '../../utils/dato';
 import { nyIsoMånedPeriode } from '../../utils/dato';
+import {
+    konverterDesimalverdiTilSkjemaVisning,
+    konverterSkjemaverdiTilDesimal,
+} from '../../utils/eøs';
 import { erBarnGyldig, erEøsPeriodeGyldig, isEmpty, isNumeric } from '../../utils/eøsValidators';
 import { useBehandling } from '../behandlingContext/BehandlingContext';
