+++ src/frontend/context/Valutakurs/ValutakursContext.tsx
@@ -1,9 +1,9 @@
 import { useEffect, useState } from 'react';
 
-import { sorterEøsPerioder } from '../../komponenter/Fagsak/Behandlingsresultat/Eøs/useEøs';
 import type { IBehandling } from '../../typer/behandling';
 import type { IRestValutakurs } from '../../typer/eøsPerioder';
 import { EøsPeriodeStatus } from '../../typer/eøsPerioder';
+import { sorterEøsPerioder } from '../../utils/eøs';
 
 interface IProps {
     åpenBehandling: IBehandling;
